converter/requirements.txt
Flask==3.0.0
requests==2.31.0
google-cloud-storage==2.13.0
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0 